requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
statsmodels>=0.14
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
openpyxl>=3.1
pyarrow>=12.0
